requirements.txt
flask
pypdf
langchain_ollama
langchain_chroma
langchain_text-splitters
langchain_community
